requirements.txt
flask
numpy
pandas
scikit-learn
python-Levenshtein
jupyter